Cast to double before dividing to avoid integer division

+++ src/lib/X6_1000.cpp
@@ -306,7 +306,7 @@ void X6_1000::set_nco_frequency(int a, int b, double freq) {
 double X6_1000::get_nco_frequency(int a, int b) {
     uint32_t phaseInc = read_dsp_register(a-1, WB_PHASE_INC_OFFSET + (b-1));
     //Undo the math in set_nco_frequency
-    return phaseInc / (1 << 18) * get_pll_frequency() / 4;
+    return static_cast<double>(phaseInc) / (1 << 18) * get_pll_frequency() / 4;
 }
 
 void X6_1000::set_threshold(int a, int b, double threshold) {
